Port Win32 SEH, Win32 file enum, and 64-bit-truncating casts
- Tactical/Handle UI.cpp: replaced __try / __except(EXCEPTION_
  EXECUTE_HANDLER) with try / catch(...) on non-MSVC. SEH catches
  hardware exceptions and is MSVC-only; standard C++ catch(...) at
  least preserves the 'don't crash' intent.
- Tactical/Handle Items.cpp:
  * Stream-extraction sites tried to operator>> into a temporary
    ((UINT16)node.item). The fields are already UINT16; dropped
    the cast so we extract into the lvalue directly.
  * get_all_files_names_within_folder() used WIN32_FIND_DATA;
    on non-Windows it now uses <dirent.h> opendir/readdir.
  * (UINT32) cast on a 64-bit pointer for opaque-handle passing
    routes through uintptr_t to silence the narrowing error. The
    receiving side (UINT32 uiApproachData) is still too small on
    LP64 -- that's a deeper bug for later.
- Tactical/Interface Dialogue.cpp: same uintptr_t intermediate for
  the pObject opaque-handle cast.
- Ja2/Options Screen.cpp: dead GetCursorPos/ScreenToClient (MousePos
  never read) removed.

+++ Tactical/Interface Dialogue.cpp
@@ -1379,7 +1379,7 @@ void CalculatePopupTextPosition( INT16 sWidth, INT16 sHeight )
 
 BOOLEAN	TalkingMenuGiveItem( UINT8 ubNPC, OBJECTTYPE *pObject, INT8 bInvPos )
 {
-	CHECKF( SpecialCharacterDialogueEvent( DIALOGUE_SPECIAL_EVENT_GIVE_ITEM, (UINT32) ubNPC, (UINT32)pObject, (UINT32) bInvPos, gTalkPanel.iFaceIndex, DIALOGUE_NPC_UI ) != FALSE );
+	CHECKF( SpecialCharacterDialogueEvent( DIALOGUE_SPECIAL_EVENT_GIVE_ITEM, (UINT32) ubNPC, (UINT32)(uintptr_t)pObject, (UINT32) bInvPos, gTalkPanel.iFaceIndex, DIALOGUE_NPC_UI ) != FALSE );
 
 	return( TRUE );
 }
